refactor comment endpoints

+++ src/comment/comment.service.ts
@@ -1,4 +1,4 @@
-import { Injectable, InternalServerErrorException } from "@nestjs/common";
+import { ConflictException, Injectable, InternalServerErrorException } from "@nestjs/common";
 import { InjectRepository } from "@nestjs/typeorm";
 import { CommentFilter, CommentRepository } from "./comment.repository";
 import { FilterCommentDto } from "./dto/filter-comment.dto";
@@ -6,6 +6,10 @@ import { PageOfDto } from "../shared/dtos/page-of.dto";
 import { CommentEntity } from "./comment.entity";
 import { extractPagination } from "../shared/utilities/extract-pagination";
 import { CreateCommentDto } from "./dto/create-comment.dto";
+import { UserDto } from "../user/dto/user.dto";
+import { UserEntity } from "../user/user.entity";
+import { prepareUser } from "../shared/utilities/prepare-user";
+import { PostEntity } from "../post/post.entity";
 
 @Injectable()
 export class CommentService {
@@ -15,12 +19,19 @@ export class CommentService {
   ) {}
 
   async getList(filterCommentDto: FilterCommentDto): Promise<PageOfDto<CommentEntity>> {
-    const { skip, take, pageNo, pageSize, comment, userId } = extractPagination(filterCommentDto);
+    const {
+      skip,
+      take,
+      pageNo,
+      pageSize,
+      comment,
+      postId
+    } = extractPagination(filterCommentDto);
 
     let commentList: CommentEntity[];
     let totalElements = 0;
 
-    let options: CommentFilter = { skip, take, userId };
+    let options: CommentFilter = { skip, take, postId };
 
     if (comment) {
       options = {
@@ -45,14 +56,21 @@ export class CommentService {
     }
   }
 
-  async create(createCommentDto: CreateCommentDto, userId: number): Promise<CommentEntity> {
+  async create(createCommentDto: CreateCommentDto, userDto: UserDto): Promise<CommentEntity> {
     const newComment: CommentEntity = new CommentEntity();
 
-    const { comment } = createCommentDto;
+    const { comment, postId } = createCommentDto;
     newComment.comment = comment;
 
+    const user: UserEntity = prepareUser(userDto);
+    const post: PostEntity = new PostEntity();
+    post.id = postId;
+
+    newComment.user = user;
+    newComment.post = post;
+
     try {
-      await newComment.save({ data: { userId } });
+      await newComment.save();
     } catch (ex) {
       throw new InternalServerErrorException(ex);
     }
@@ -60,8 +78,13 @@ export class CommentService {
     return newComment;
   }
 
-  async delete(id: number): Promise<CommentEntity> {
+  async delete(id: number, userId: number): Promise<CommentEntity> {
     const foundComment: CommentEntity = await this.commentRepository.findOne(id);
+
+    if (foundComment.user.id !== userId) throw new ConflictException({
+      data: null,
+      result: 'You can not delete comment made by someone else!'
+    });
 
     try {
       await this.commentRepository.remove(foundComment);
